* skk-tut.el (skktut-init-variables-alist): Remove `(skk-convert-okurigana-into-katakana . nil)' element.
* skk-vars.el (skk-convert-okurigana-into-katakana): Removed.

* skk.el (skk-start-henkan): When SKK-KATAKANA is non-nil and ARG
is equal to 1, not convert a string of a region to hiragana but
convert a string in SKK-HENKAN-KEY itself.
(skk-set-okurigana): Ditto.  Also convert SKK-HENKAN-OKURIGANA.
(skk-hiragana-to-katakana): Bug fixed.
(skk-katakana-to-hiragana): Ditto.
(skk-toggle-characters): Convert region even if SKK-HENKAN-ACTIVE
is non-nil.  Try to recognize characters type by checking second
characters and more even if type of first character is `unknown'.
Doc updated.
(skk-previous-candidate): Put out katakana when SKK-KATAKANA is non-nil.
(skk-kakutei-cleanup-buffer): Not to convert okurigana into
katakana.