Keep bundled packages of the load-path
It is not a good idea to bundle third-party packages because such
copies tend to not get updated.  Users are then at risk of loading
the outdated bundled copy instead of the upstream version that they
also have installed.

This commit mitigates that issue somewhat by placing the bundled
libraries in dedicated directories.  These directories contain a file
named ".nosearch", which makes it less likely that they will end up
on the `load-path'.  Additionally this commit updates some `require'
forms to load the bundled versions if necessary.