Delete bundled copy of vip.el
It is not a good idea to bundle third-party packages because such
copies tend to not get updated.  Users are then at risk of loading
the outdated bundled copy instead of the upstream version that they
also have installed.

It does not appear that vip.el is actually used by this package, so
this commit simply removes it.