Add library matching package name
Melpa extracts metadata from the library whose name matches the name
of the package.  If that file does not exist, then the metadata
(including such essential things as a description of the package) is
missing when the user tries to display information about the package
using package.el.

For backward compatibility just add an additional file ddskk.el, with
the same metadata as skk.el.  Provide ddskk and require skk in the new
file.  That way existing users can continue to load the package as
they used to do (e.g., by requiring skk), but it also becomes possible
to treat ddskk.el as the new entry point.

Renaming the *package* to "skk" would also have been an option but
because other packages already depend on it as "ddskk", that seems
like a worse option.